seperate plot_options into presenter class Re #107
